requirements.txt
# Application
click==8.1.7
pyyaml==6.0.2
rich==13.9.4

# Testing
pytest==8.3.4
